Improved scrolling on About Dialog
Updated database so card number is a Text (for transform's a/b). Fixed MtgCard, JsonCardParser for this too.

Updated cardview, resultlist, CardDbAdapter to add a transform button and code for applicable cards.

Added a SharedPreference for DB version. The DB shouldn't get updated often and has to come through the APK. Implemented in main.