Added support for Android 1.5 and high/medium resolution screens. Cupcake requires a /res/drawable folder, so thats a copy of the -mdpi-v4 folder. The -v4 is because the internet told me to.
Disabling the animations for transformations is commented out, as it is in API level 5.

The official Google method for multiple resolutions in API level 3 creates a warning, so there goes the warningless project. (targetSdkVersion > minSdsVersion in manifest)

Cleaned up warnings in RNG.java

Put thread operation in a giant try/catch block for debugging. I think that one person had issues with his phone though, not the app.