changed package name to a more acceptable convention, although I don't own gelakinetic.com. As of now, no one does.
Changed soft keyboard popups on the search page to have a "search" button instead of an "enter" button. No more newlines!

Recycled bitmap in cardview to prevent memory leaks.
Added the very basic deck management beginnnings. The button to access it from the main page is set to View.GONE. Get rid of that to write/debug it. THIS IS NOT REMOTELY USER READY.