Issue 29 - This was fixed earlier. Can we close it?
Issue 32 - Changed the HorizontalListView to one from a fork of AndroidDev-Lib. ListAdapters work now! All it needs is UI polish. Lots of it...

Issue 59 - Cleaned up query and moved off UI thread. Should be good to close.

Issue 60 - Added a button in CardView's menu to quit to search. It works, and as a warning, each activity's onResume now has to clear a state int so it won't quit forever

Changed version code / number in manifest. Changed "What's new" string for v1.3. Added props to Alex in "about"