Added thank you for dinocore1's horizontalListView
Added breakpoints to the what's new dialog

Fixed a bug relating to replacing the DB from the APK when the app is updated. This was because the contentProvider for autocomplete opened the db before the regular check could occur in onCreate in mainActivity

Added eclipse config files. Mike, you can deal with it. Without them, the project won't work in Eclipse w/o extra configs (confusing to new devs). If you want to put your ANT script in here too, thats fine.

Added code to calculate the unix time of a set's release in MtgXml2Json.