Fixed an FC in CardTrading where if there was an active fetch which did not finish, and another fetch did finish, it would die when looking for a "price" which wasn't populated yet.
Fixed a bug in FC where if you searched for a card which failed the fetch "i.e. Alpha Black Lotus" and then changed the set to one with a price "i.e. Unlimited Black Lotus" there error flag wouldn't clear.

The trader now automatically removes entries for which the card does not exist, or could not be found in the database. It leaves in card where the fetch failed, otherwise Black Lotus would never be added (since the default set is Alpha, and the fetch fails). It also displays a toast