Merged Alex's and my versions of rules activity layouting.  I went back to multiple layouts for the rows because it a. allows for what I see as a cleaner difference between the levels and b. separates layout from implementation.  It also allows for much nicer layout when a rule name spills over onto another line vs the one-layout solution.  I actually really like some unintentional side-affects of our fusion, including the way the glossary entry at the bottom of the main list looks.
Also, both of our sets of fixes are now in effect, like Alex's improved example
formatting and capture of search button presses, and my removal of the annoying
blink when clicking non-link text in rules.

The lesson I've learned here is not to muck with code someone else might be
working on!  I'll communicate better going forward