Renamed the unfortunate _ActionBarSherlockTrojanHorse to _ActionBarSherlockTH. It's not a virus, I swear!
Added ActionBar and search key to preferences activity.

Fixed bug where you could cancel an uncancelable dialog with the search key (thanks April :P)

Moved OTA code to FamiliarActivity. This is to protect against the off chance that the user launches the app with the widget before opening the main page after instalation. Fixed the asyncTask to fire off after checking the preference, rather than checking the preference in the task

Moved mCtx, mDbHelper, and preferences to FamiliarActivity. That should clean things up a bit. mDbHelper is opened/closed in onCreate()/onDestroy() in FamiliarActivity. Preferences and mCtx are instantiated in onCreate()

Formatting

Commented out pieces which broke because of a partial commit. Please don't do that in the future. Try to only commit code which compiles