Made a helper class for fetchPriceTask and cardData which is shared between CardTradingActivity and WishlistActivity.
Removed stuff about the saved instance state, since autosave takes care of it.

Fixed wishlist verbose mode not working. Protip: if you set a view's visibility to View.GONE and you might want to display it later, you have to set the visibility to View.VISIBLE. It was being set to View.GONE with the activity opened and the list was filled, but not being set to View.VISIBLE when the information returned from the asynctask.