Fixed warning in ActionBarSherlock, NPlayerLifeActivity, SearchActivity
Commented out debug code in ResultListActivity

Removed QueuedAsyncTask, replaced with LinkedList of pending tasks and ArrayList of active tasks. Whenever a task is added, or a task finishes, it will look to see if there is enough room for another task to be executed, and execute it. If the tasks are canceled, the pending queue will be cleared and all currently running tasks will not try to start another when they finish. This affected both CardTradingActivity and WishlistActivity.