TL;DR Added tablet optimized card search. Changes shouldn't be noticeable on phones.
Added three-pane layout which is automatically used on tablets (see refs.xml).
MainActivity has a set of functions to show and hide panes, add fragments to each pane, and send bundles between the fragments.
A InFragmentMenuLoader now inflates options menus into a split actionbar-like menu. This is inflated in the fragment, so the different menus do not collide in the actionbar when displaying multiple fragments. Each fragment using this kind of menu must have an enclosing LinearLayout called master_layout with no margins. The menu is a LinearLayout which gets added to master_layout, so the rest of the UI elements should have weight=1 and height=0dp
Removed random menu option from SearchViewFragment, because it exists in the ResultListFragment menu
Added holo_blue_dark and renamed holo_blue to holo_blue_light.

Things not working:
Scroll buttons in the cardview in three-pane mode

Things missing:
Icons for all CardView menu options